+++ tests/modules/test_pbench.py
@@ -1,10 +1,8 @@
 import io
-import itertools
 import logging
 import math
 import os
 import pprint
-import random
 import time
 
 import urwid
@@ -324,7 +322,7 @@ if not is_windows():
             obj.pbench.generate_payload(obj.get_scenario())
             payload_count = len(obj.get_scenario().get('requests', []))
             sch = Scheduler(load, open(obj.pbench.payload_file, 'rb'), logging.getLogger(''))
-            estimated_schedule_size = obj.pbench._estimate_max_progress(load, payload_count)
+            estimated_schedule_size = obj.pbench._estimate_schedule_size(load, payload_count)
             logging.debug("Estimated schedule size: %s", estimated_schedule_size)
             items = list(sch.generate())
             actual_schedule_size = len(items)